add link to player details page
