+++ rel_bball_analytics/templates/players/details.html
@@ -1,0 +1,10 @@
+{% extends 'base.html' %}
+
+{% block content %}
+<h1 class="mt-4 mb-2">{{ player_data.firstname }} {{ player_data.lastname }}</h1>
+<div class="container mt-4">
+    <!-- TODO: Display data properly -->
+    {{ player_data }}
+</div>
+
+{% endblock %}
